test: move tests into dedicated directory and only test the public interface
