+++ .gitignore
@@ -17,6 +17,7 @@
 !Cargo.toml
 !Cargo.lock
 !src/*.rs
+!tests/*.rs
 !build.rs
 target/**/*
