+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod analyzer;
+pub mod git;
+pub mod markdown;
